+++ src/main.ts
@@ -4,7 +4,7 @@ import { ConfigService } from '@nestjs/config';
 import { Logger, ValidationPipe } from '@nestjs/common';
 import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 import { AppModule } from '@/app.module';
-import { HttpExceptionFilter } from '@/infra/http/nest/config/http-filter-exception';
+import { HttpExceptionFilter } from '@/infra/http/nest/@config/http-filter-exception';
 
 async function bootstrap() {
   /* Set config initial */
